Refactor import paths in D-FINE module and clean up unused code
