+++ src/otx/algo/detection/d_fine.py
@@ -16,7 +16,7 @@ from torchvision.tv_tensors import BoundingBoxFormat
 
 from otx.algo.detection.backbones.hgnetv2 import HGNetv2
 from otx.algo.detection.detectors import DETR
-from otx.algo.detection.detectors.dfine.dfine_decoder import DFINETransformer
+from otx.algo.detection.heads.dfine_decoder import DFINETransformer
 from otx.algo.detection.losses.dfine_loss import DFINECriterion
 from otx.algo.detection.necks.dfine_hybrid_encoder import HybridEncoder
 from otx.algo.utils.mmengine_utils import load_checkpoint
